+++ editors/vscode/CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to the "swarm-language" extension will be documented in this
 
 ## WIP
 - [Highlighter] added `swap` command
+- [Highlighter] added `split` and `chars` functions
 
 ## version 0.0.6
 - [Highlighter] **FIX:** the number pattern now matches properly, previously it broke everything!
@@ -33,4 +34,4 @@ All notable changes to the "swarm-language" extension will be documented in this
 
 ## version 0.0.1
 
-- Initial release with support for higlighting
+- Initial release with support for highlighting